requirements.txt
streamlit
openai
weaviate-client
azure-ai-documentintelligence
azure-core
PyYAML